requirements.txt
Django==4.0.3
django-dry-rest-permissions==1.2.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
Markdown==3.3.6
django-cors-headers
